IMPALA-9250: catch exceptions of orc::RowReader::createRowBatch
The ORC lib uses exceptions to report failures. We are missing
exception handling in invoking orc::RowReader::createRowBatch which
requires memory allocation and could raise exceptions when it fails.
This patch simply adds a catch clause for it.

To simplify the codes, a macro is added for catching the ORC exceptions
with given message and an additional statement.

Tests:
 - TODO: run test_scanner_fuzz.py 100 times

Change-Id: I76e36a238220e7bed1cbbdcb3fc7d35394bfa023