IMPALA-11857: Connect join build fragment to join in graphical plan
When support was added for join build sink, the plan JSON and plan
rendering logic were not updated to handle the new sink type. As a
result, the linkage from the join build fragment to its corresponding
join node were not expressed in the JSON and build fragments nodes were
rendered as orphaned.

This change adds a new JSON join_build_target field to join build
fragments and connects the build fragment to the join with a green dashed
line similar to the red dashed line used for data sender fragments.

Also changed the SVG fill type to 'none' for exchange edges to avoid
rendering a black triangle if the right child was an exchange as in the
join build case.

Testing:
- Manual testing with TPCH queries and reviewing rendered plan diagrams

Change-Id: I80af977e5c5e869268d3ee68fafe541adadc239d
Reviewed-on: http://gerrit.cloudera.org:8080/19437
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>