IMPALA-12460: Add lag and histogram of event processing in the log
This patch logs the lag of the event processing which is
(latestEventTime - lastSyncedEventTime) at the end of processing an
event batch. If the batch is slow to process, we also log the top-10
expensive events and top-10 tables contribute to this. Admins can decide
whether to disable event processing on these tables.

Also adds a thread annotation on the event-processor thread so in
jstacks, we know what event is under processing.

Change-Id: Ib9421b5e26bfa2324217ec9695fbd81636727d22