IMPALA-12788: Fix HBaseTable still get loaded even if HBase is down
When loading a table backed by HBase, it's intended to check whether the
table exists in HBase in HBaseTable.load() and loadFromThrift().
However, the current check just gets the table object and then closes
it. This won't fail even if HBase is down. See JIRA description for the
stacktrace.

This patch fixes the check to fetch the column family names which is a
light-weight request and will fail if HBase is down or the table doesn't
exist in HBase.

Splits TestNestedStructsInSelectList.test_struct_in_select_list and
TestDdlStatements.test_alter_set_column_stats to skip the HBase part
when running on S3.

Tests:
 - Run CORE tests on S3

Change-Id: Ib497f11ecc338d0f84d3d7bd8ccfcf8da4def0cb