WIP IMPALA-12933: Avoid fetching unneccessary event types
There are several places that catalogd will fetch all events of a
specifit type on a table. E.g. in TableLoader#load(), if the table has
an old createEventId, catalogd will fetch all CREATE_TABLE events after
that createEventId on the table.

Fetching the list of events is expensive since the filtering is done on
client side, i.e. catalogd fetches all events and filter them locally
based on the event type and table name. This could take hours if there
are lots of events (e.g 1M) in HMS.

This patch sets the eventTypeSkipList with the complement set of the
wanted type. So the get_next_notification RPC can filter out some events
on HMS side.

Also adds UPDATE_PART_COL_STAT_EVENT to the default skip list.

Change-Id: Ieabe714328aa2cc605cb62b85ae8aa4bd537dbe9