IMPALA-13117: (part-1) simplify listWithLocations in FileMetadataLoader
There are two HDFS APIs to list file metadata of a folder - listFiles
and listStatus. The first one returns a list of LocatedFileStatus which
has the info of block locations. The second one returns a list of
FileStatus without block locations. listStatus is usually used on
storages that don't have blocks, e.g. S3, or used in scenarios that we
want to list basic info (name, length, last modified time) reuse
existing file descriptors.

This patch simplifies the check on whether to list files with block
locations. When the table/partition is marked as HDFS caching enabled,
or if there are no existing file descriptors to reuse, we invoke the
listFiles API.

Also tracks whether there are file changes in FileMetadataLoader to
speed up the normal case in hasFilesChangedCompareTo().

Tests
 - Ran CORE tests

Change-Id: Ib690930b57a161c98f07ccc2304860b788e9cd9f