IMPALA-11402: Add limit on files fetched by getPartialCatalogObject requests
Adds a new flag, catalog_partial_fetch_max_files, to define the max
number of file descriptors allowed in a response of
getPartialCatalogObject. Catalogd will truncate the response in
partition level when it's too big. In this case, some partitions will be
missing. Coordinator should send new requests to fetch the missing
partitions.

Adds a new optional field, allow_incomplete_data, in TTableInfoSelector
for catalogd to avoid breaking coordinators in old versions. It should
always be set to true. Coordinators in old versions send the requests
without that field so it's false in catalogd's perspective.

Change-Id: Ibb13fec20de5a17e7fc33613ca5cdebb9ac1a1e5