IMPALA-13684: Improve waitForHmsEvent() to only wait for related events
waitForHmsEvent is a catalogd RPC for coordinators to send a requested
db/table names to catalogd and wait until it's safe (i.e. no stale
metadata) to start analyze the statement. Currently, catalogd waits
until it syncs to the latest HMS event regardless what the query is.
This patch improves this by only checking related events. In the ideal
case, if there are no pending events that are related, query doesn't
need to wait.

We check the pending events in the following way:
 - For queries that need the db list, i.e. SHOW DATABASES, check pending
   CREATE_DATABASE, DROP_DATABASE events on all dbs.
 - For db statements like SHOW FUNCTIONS, CREATE/ALTER/DROP DATABASE,
   check pending CREATE/ALTER/DROP events on that db. For db names that
   are used in the query and are missing in catalog, also check their
   events.
   - For db statements that require the table list, i.e. SHOW TABLES,
     also check CREATE_TABLE, DROP_TABLE events under that db.
 - For loaded transactional tables, check all the pending COMMIT_TXN,
   ABORT_TXN events.
 - For all the table names,
   - if they are all missing/unloaded in the catalog, check all the
     pending CREATE_TABLE, DROP_TABLE events on them for their
     existence.
   - Otherwise, some of them are loaded, check all the table events on
     them.

This patch leverages the HMS API to fetch events of several tables under
the same db in batch. MetastoreEventsProcessor.MetaDataFilter is
improved for this.

Tests:
 - Ran metadata/test_event_processing.py::TestEventSyncWaiting and
   custom_cluster/test_kudu.py::TestKuduHMSIntegration::test_drop_managed_kudu_table

Change-Id: Ic033b7e197cd19505653c3ff80c4857cc474bcfc