IMPALA-13884: Improve catalogd logs
Add more logs in HMS event processing to print the target of the event.
For AlterPartitionEvent, the partition name will also be printed. Note
that Add/DropPartitionEvent could have multiple partitions so ignore
them for now.

In the reason of loading table metadata, adds the event id.

For logs of execDdl requests on AlterTable operations, add the
AlterType.

Adds a log when start recovering partitions.

Also adds an overload of AcidUtils.isTransactionalTable() for FeTable to
simplify some codes.

Change-Id: I204d5922e055fd8501b5573e3b913f8874d891d6