IMPALA-13285: Ignore COMMIT_TXN events on Apache Hive 3
In Apache Hive 3, HMS doesn't provide the API to retrive WriteEvents
info of a given transaction. COMMIT_TXN event just contains a
transaction id so Impala can't process it.

This patch ignores COMMIT_TXN events when building on Apache Hive 3.
Some tests in MetastoreEventsProcessorTest and EventExecutorServiceTest
are skipped due to this.

Tests:
 - Manually tested on Apache Hive 3. Verified EventProcessor still
   working after receiving COMMIT_TXN events.
 - Passed some tests in MetastoreEventsProcessorTest and
   EventExecutorServiceTest that previously failed due to EventProcessor
   goes into ERROR state.

Change-Id: I863e39b3702028a14e83fed1fe912b441f2c28db